// packages/mobile/modules/state.d.ts
/*
 Generated by typeshare 1.3.0
*/

export interface CollectionState {
	id: number;
	library: number;
	title: string;
	items?: number[];
	thumbnail?: ThumbnailState;
}

export interface PlaylistState {
	id: number;
	title: string;
	videos?: number[];
}

export interface LibraryState {
	id: number;
	title: string;
	type: LibraryType;
}

export interface SeasonState {
	id: number;
	show: number;
	index: number;
	title: string;
}

export interface ShowState {
	id: number;
	library: number;
	title: string;
	year: number;
	thumbnail?: ThumbnailState;
}

export interface MovieState {
	library: number;
	year: number;
}

export interface EpisodeState {
	season: number;
	index: number;
}

export interface VideoState {
	id: number;
	title: string;
	detail: VideoDetail;
	thumbnail?: ThumbnailState;
	download?: DownloadState;
}

export interface ServerState {
	token?: string;
	name: string;
	playlists?: PlaylistState[];
	collections?: CollectionState[];
	libraries?: LibraryState[];
	shows?: ShowState[];
	seasons?: SeasonState[];
	videos?: VideoState[];
}

export interface State {
	clientId: string;
	servers?: Record<string, ServerState>;
}